save a backlog of attachments too
This will help us afterwards if we label a defined state, covering exact
object-item, object-attachment and object-achievement.

If we do no save the attachment we can later not reference to the
particular point.

Missing piece: currently we ALWAYS update the achiement (add to the end
of the list). Later we should compare all values and only if there are a
difference we should update the entry.

One crucial point: the compare operation MUST ignore the submitter -
because they will change permanently.

Signed-off-by: Hagen Paul Pfeifer <hagen@jauu.net>