report list with reference to report
The Report part is now devided by tabs into a section where all current existing reports are listed and a section where a new report can be created.
The reports are listed by date and are referenced to the report document, so by clicking on the report name the document is displayed.

Signed-off-by: Alexander Sarmanow <alexsar@gmx.de>