introduce snippet_db.py
When a snippet is registrated in mime_data_db.py while an object is posted, snippet_db.py is saving the base64 encoded snippet in a tmp file to execute it later, when posting process or similiar is completed.
After an successful execution the equal meta data in subcontainer or achievement gets added by the outcame of the snippet.

Signed-off-by: Alexander Sarmanow <alexsar@gmx.de>