higher preference for snippet mimetype
The problem was that in case of same data (sha) but different mimetypes (e.g. 'snippet-x-...' and 'text/plain') only one of them is listed in attr.db. So it is possible that one is more important (in this example snippet), but is overwritten by text/plain. Because of that the snippet-images couldn't be loaded.

Signed-off-by: Alexander Sarmanow <alexsar@gmx.de>